use `f32` instead of `f64`

+++ autd3-driver/src/acoustics/directivity/mod.rs
@@ -7,8 +7,8 @@ pub use sphere::Sphere;
 pub use t4010a1::T4010A1;
 
 pub trait Directivity: Send + Sync {
-    fn directivity(theta_deg: f64) -> f64;
-    fn directivity_from_dir(axial_direction: &Vector3, target: &Vector3) -> f64 {
+    fn directivity(theta_deg: f32) -> f32;
+    fn directivity_from_dir(axial_direction: &Vector3, target: &Vector3) -> f32 {
         Self::directivity(
             (axial_direction.cross(target).norm())
                 .atan2(axial_direction.dot(target))
@@ -24,7 +24,7 @@ pub mod tests {
     pub struct TestDirectivity {}
 
     impl Directivity for TestDirectivity {
-        fn directivity(t: f64) -> f64 {
+        fn directivity(t: f32) -> f32 {
             t
         }
     }
@@ -35,7 +35,7 @@ pub mod tests {
     #[case::dir_y(90., Vector3::y(), Vector3::z())]
     #[case::dir_z(0., Vector3::z(), Vector3::z())]
     fn test_directivity_from_dir(
-        #[case] expected: f64,
+        #[case] expected: f32,
         #[case] target: Vector3,
         #[case] dir: Vector3,
     ) {
