+++ autd3-driver/src/acoustics/directivity/sphere.rs
@@ -4,12 +4,12 @@ pub struct Sphere {}
 
 impl Directivity for Sphere {
     #[inline]
-    fn directivity(_: f64) -> f64 {
+    fn directivity(_: f32) -> f32 {
         1.
     }
 
     #[inline]
-    fn directivity_from_dir(_: &Vector3, _: &Vector3) -> f64 {
+    fn directivity_from_dir(_: &Vector3, _: &Vector3) -> f32 {
         1.
     }
 }
@@ -31,7 +31,7 @@ mod tests {
     #[case::dir_x(1., Vector3::x())]
     #[case::dir_y(1., Vector3::y())]
     #[case::dir_z(1., Vector3::z())]
-    fn test_directivity_sphere_from_dir(#[case] expected: f64, #[case] target: Vector3) {
+    fn test_directivity_sphere_from_dir(#[case] expected: f32, #[case] target: Vector3) {
         let mut rng = rand::thread_rng();
         let dir = Vector3::new(rng.gen(), rng.gen(), rng.gen());
         assert_eq!(expected, Sphere::directivity_from_dir(&dir, &target));
